Direct Multistep Forecast

Make forecast using a horizon-specific estimated OLS model.

fcast(df, lags, .h, ic = BIC, .var = "all", .diff = FALSE,
  wav = FALSE, xreg = NULL, ...)

Arguments

df
A data frame. The first column of the data frame must be the base variable for constructing the dependent variable, the multiperiod ahead value to be forecasted, \(x_{t + h}\). If wav = TRUE, ncol of df can be equal to one (just the dependent variable), otherwise, ncol(df) must be greater than one, even if .var = "ar".

lags
An integer vector defining the lags of the regressors. If wav = FALSE, the length of the vector lags have to be the same as the number of columns in df. However, if wav = TRUE, an additional element in lags must be add such that the last element in the vector lags is the lag associatd with the wavelet regressor.

.h
\(h\) step-ahead.

ic
Information criterion, BIC or AIC. When searching for the best model the dataset is adjusted so that every model have the same data length for appropriate comparasion.

.var
A string to determine how the model will be specificated: "all" (default), "ar" or "ar_out".

.diff
Logical FALSE or TRUE. If TRUE, the dependent variable is differentiated. See wavdrcast-package for the implication on the model specification and wavelet variable.

wav
Logical. If TRUE, a wavelet-based signal is add to df, where the signal is related to the explained variable, the first column of df.

xreg
Data frame. Exogeunous variable not subjected to be lagged. The number of rows must be the same as in df.

...
Further arguments passed to wavsigmap::signal.

Value

A double vector with forecasts from 1 to \(h\) step-ahead.

Details

See wavdrcast-package and model.

See also

Examples

fcast(df = inf[1], lags = c(2, 1), .h = 12, wav = TRUE)

#>          1          2          3          4          5          6          7 
#> 0.08408635 0.10075261 0.21604766 0.28420205 0.32169176 0.30622540 0.22984749 
#>          8          9         10         11         12 
#> 0.17161654 0.18542831 0.15362946 0.17018959 0.20119350 

fcast(df = gdp[1], lags = c(2, 1), .h = 12, wav = TRUE, .diff = TRUE)

#>        1        2        3        4        5        6        7        8 
#> 9.869637 9.875522 9.884631 9.896126 9.904314 9.908513 9.914665 9.920262 
#>        9       10       11       12 
#> 9.925948 9.935739 9.937006 9.927519 

fcast(df = gdp, lags = c(2, 1), .h = 12, .diff = TRUE)

#>        1        2        3        4        5        6        7        8 
#> 9.828319 9.831948 9.835816 9.840032 9.845794 9.850865 9.856438 9.863763 
#>        9       10       11       12 
#> 9.868739 9.873068 9.881494 9.886520