Direct Model

This function estimates a direct least square horizon-specific model.

model(df, lags, .h, ic = BIC, .var = "all", .diff = FALSE,
  wav = FALSE, xreg = NULL, ...)

Arguments

df
A data frame. The first column of the data frame must be the base variable for constructing the dependent variable, the multiperiod ahead value to be forecasted, \(x_{t + h}\). If wav = TRUE, ncol of df can be equal to one (just the dependent variable), otherwise, ncol(df) must be greater than one, even if .var = "ar".

lags
An integer vector defining the lags of the regressors. If wav = FALSE, the length of the vector lags have to be the same as the number of columns in df. However, if wav = TRUE, an additional element in lags must be add such that the last element in the vector lags is the lag associatd with the wavelet regressor.

.h
\(h\) step-ahead.

ic
Information criterion, BIC or AIC. When searching for the best model the dataset is adjusted so that every model have the same data length for appropriate comparasion.

.var
A string to determine how the model will be specificated: "all" (default), "ar" or "ar_out".

.diff
Logical FALSE or TRUE. If TRUE, the dependent variable is differentiated. See wavdrcast-package for the implication on the model specification and wavelet variable.

wav
Logical. If TRUE, a wavelet-based signal is add to df, where the signal is related to the explained variable, the first column of df.

xreg
Data frame. Exogeunous variable not subjected to be lagged. The number of rows must be the same as in df.

...
Further arguments passed to wavsigmap::signal.

Value

A list with the following elements:

xreg

A tibble with the data regressors used to run the model choosed by the information criterion. This can be a different data set from that one returned by the lm object.

model

An object of class lm.

Details

General OLS regression model is writting as:

\(x_{t + h} = \alpha + \mathbf{{\beta}}'\mathbf{x}_t + \mathbf{{\gamma}}'\mathbf{z}_t + \mathbf{{\theta}}'\mathbf{d}_t + \epsilon_{t + h}\),

where variables are explained in wavdrcast-package details section. Arguments .var, .diff and wav allow one to estimate subsets of this general model, as discussed in wavdrcast-package. If .var = "all", the general model is estimated (unless xreg = NULL, so that \(\mathbf{d}\) is out). If .var = "ar", \(\mathbf{z}\) terms are excluded (not \(\mathbf{d}\), if included). Finally, if .var = "ar_out", the model is estimated without \(\mathbf{x}\) regressors.

See also

Examples

model(df = inf, lags = c(2, 0), .h = 2, .var = "ar")

#> $xreg
#> # A tibble: 598 x 1
#>    cpi_lag0
#>       <dbl>
#>  1    0.265
#>  2    0.529
#>  3    0.526
#>  4    0.785
#>  5    0.260
#>  6    0.518
#>  7    0.515
#>  8    0    
#>  9    0.513
#> 10    0.510
#> # ... with 588 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = cpi_lead2 ~ cpi_lag0)
#> 
#> Coefficients:
#> (Intercept)     cpi_lag0  
#>      0.1985       0.3828  
#> 
#> 

model(df = inf, lags = c(2, 2), .h = 2)

#> $xreg
#> # A tibble: 598 x 4
#>    cpi_lag0 core_lag0 core_lag1 core_lag2
#>       <dbl>     <dbl>     <dbl>     <dbl>
#>  1    0.265     0.253    NA        NA    
#>  2    0.529     0.505     0.253    NA    
#>  3    0.526     0.754     0.505     0.253
#>  4    0.785     0.499     0.754     0.505
#>  5    0.260     0.496     0.499     0.754
#>  6    0.518     0.741     0.496     0.499
#>  7    0.515     0.245     0.741     0.496
#>  8    0         0.244     0.245     0.741
#>  9    0.513     0.732     0.244     0.245
#> 10    0.510     0.726     0.732     0.244
#> # ... with 588 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = cpi_lead2 ~ cpi_lag0 + core_lag0 + core_lag1 + 
#>     core_lag2)
#> 
#> Coefficients:
#> (Intercept)     cpi_lag0    core_lag0    core_lag1    core_lag2  
#>      0.1178       0.2155       0.1355       0.1084       0.1737  
#> 
#> 

model(df = inf[1], lags = c(2, 2), .h = 2, wav = TRUE)

#> $xreg
#> # A tibble: 598 x 3
#>    cpi_lag0 cpi_lag1 cpi_wav_lag0
#>       <dbl>    <dbl>        <dbl>
#>  1    0.265   NA            0.451
#>  2    0.529    0.265        0.451
#>  3    0.526    0.529        0.451
#>  4    0.785    0.526        0.450
#>  5    0.260    0.785        0.447
#>  6    0.518    0.260        0.444
#>  7    0.515    0.518        0.439
#>  8    0        0.515        0.434
#>  9    0.513    0            0.427
#> 10    0.510    0.513        0.419
#> # ... with 588 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = cpi_lead2 ~ cpi_lag0 + cpi_lag1 + cpi_wav_lag0)
#> 
#> Coefficients:
#>  (Intercept)      cpi_lag0      cpi_lag1  cpi_wav_lag0  
#>      0.08663      -0.13922      -0.14991       1.01753  
#> 
#> 

model(df = gdp, lags = c(2, 2), .h = 2, .diff = TRUE)

#> $xreg
#> # A tibble: 282 x 3
#>    lgdp_diff_lag0  gap_lag0  gap_lag1
#>             <dbl>     <dbl>     <dbl>
#>  1       NA       -0.000635 NA       
#>  2       -0.00342 -0.0167   -0.000635
#>  3        0.0103  -0.0191   -0.0167  
#>  4       -0.00843 -0.0403   -0.0191  
#>  5        0.0385  -0.0146   -0.0403  
#>  6        0.0300   0.00240  -0.0146  
#>  7        0.0379   0.0271    0.00240 
#>  8        0.0190   0.0328    0.0271  
#>  9        0.0135   0.0326    0.0328  
#> 10        0.0172   0.0355    0.0326  
#> # ... with 272 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = lgdp_diff_lead2 ~ lgdp_diff_lag0 + gap_lag0 + 
#>     gap_lag1)
#> 
#> Coefficients:
#>    (Intercept)  lgdp_diff_lag0        gap_lag0        gap_lag1  
#>        -0.0108          3.1714         -2.8187          2.5336  
#> 
#> 

model(df = gdp[1], lags = c(2, 2), .h = 2, .diff = TRUE, wav = TRUE)

#> $xreg
#> # A tibble: 282 x 2
#>    lgdp_diff_lag0 lgdp_wav_diff_lag0
#>             <dbl>              <dbl>
#>  1       NA                 0.00313 
#>  2       -0.00342          -0.00176 
#>  3        0.0103            0.00400 
#>  4       -0.00843          -0.00278 
#>  5        0.0385           -0.000862
#>  6        0.0300           -0.00563 
#>  7        0.0379            0.00344 
#>  8        0.0190            0.00276 
#>  9        0.0135           -0.000951
#> 10        0.0172           -0.000257
#> # ... with 272 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = lgdp_diff_lead2 ~ lgdp_diff_lag0 + lgdp_wav_diff_lag0)
#> 
#> Coefficients:
#>        (Intercept)      lgdp_diff_lag0  lgdp_wav_diff_lag0  
#>           0.007348            1.062551           -3.157724  
#> 
#> 

model(df = gdp[1], lags = c(0, 0), .h = 2, .diff = TRUE, wav = TRUE, .var = "ar_out")

#> $xreg
#> # A tibble: 282 x 1
#>    lgdp_wav_diff_lag0
#>                 <dbl>
#>  1           0.00313 
#>  2          -0.00176 
#>  3           0.00400 
#>  4          -0.00278 
#>  5          -0.000862
#>  6          -0.00563 
#>  7           0.00344 
#>  8           0.00276 
#>  9          -0.000951
#> 10          -0.000257
#> # ... with 272 more rows
#> 
#> $model
#> 
#> Call:
#> stats::lm(formula = lgdp_diff_lead2 ~ lgdp_wav_diff_lag0)
#> 
#> Coefficients:
#>        (Intercept)  lgdp_wav_diff_lag0  
#>             0.0157             -1.7799  
#> 
#>